requirements.txt
djangorestframework
psycopg2-binary
requests
python-dotenv
django-redis
redis
celery
base58

